Move build flags to project-side nim.cfg and support more micros
Previously, the nim compiler build flags were in the Python script.
Since command-line settings prevent the use of any config file settings,
this tool would overwrite a user's build configuration.  We should avoid this.

This commit moves the build flags to a new nim.cfg file that
is copied to the user's project folder if one does not exist.
This allows the user to change the build settings as needed.
Two build flags, path and cpu, must be dynamically written into the config file
because they come from the platformio values in the Python script.

I added the "opt:size" build flag to nim.cfg.

I also added a support for more microcontroller families.
A lookup table is used to convert the $PIOPLATFORM
variable to the CPU type used by the Nim compiler.
The result is put in the "cpu" build flag in the config file.